fix: address PRRT_kwDOSJAM6s6RRf2l — split oversized hosted payload tests
